+++ src/.vuepress/sidebar.js
@@ -55,7 +55,7 @@ function getViewerSidebar() {
         "examples/ifc_annotations",
         "examples/plan_annotations",
         "examples/global_components",
-        "examples/filter_elements",
+        "examples/partial_loading",
       ],
     },
     {