+++ src/.vuepress/viewer/index.js
@@ -1,6 +1,6 @@
 import contextMenu from "./config/contextMenu.js";
 import contextPlugins from "./config/contextPlugins.js";
-import filterElements from "./config/filterElements.js";
+import partialLoading from "./config/partialLoading.js";
 import globalComponents from "./config/globalComponents.js";
 import guiLayout from "./config/guiLayout.js";
 import hubs from "./config/hubs.js";
@@ -16,7 +16,7 @@ import windowUI from "./config/windowUI.js";
 const configs = new Map([
   ["contextMenu", contextMenu],
   ["contextPlugins", contextPlugins],
-  ["filterElements", filterElements],
+  ["partialLoading", partialLoading],
   ["globalComponents", globalComponents],
   ["guiLayout", guiLayout],
   ["hubs", hubs],